[Do Not Pull] [Not working] Convert SQL to KVP
Hi,

Do not merge this pull request as it'll break the resource. I've converted the SQL functions to use KVP with the intention of making the resource truly standalone. However, these changes seem to screw up the callbacks. Throwing up erros on these lines:

[    script:inventory] SCRIPT ERROR: @inventory/main/callbacks.lua:24: attempt to call a nil value (global 'TriggerServerEvent')
[    script:inventory] > TriggerCallback (@inventory/main/callbacks.lua:24)
[    script:inventory] > GetPlayer (@inventory/main/player.lua:159)
[    script:inventory] > handler (@inventory/main/events.lua:6)

Any idea whats causing this? I'll submit a PR with a config to make it optional to use sql or kvp once I can get past this!